Update to stateNode, rpiLightingNode, and ilocator_pythonLib
Changes to stateNode to make it functional for testing on Nov 9, as well as some cleanup.

Updated version of the ilocator_pythonLib.py file

No notable changes to rpiLightingNode, this is next up on the todo list if stateNode is good to go.

Also now putting the entire catkin workspace in this folder, can revert if needed.